Trait legion::borrow::UnsafeClone

pub trait UnsafeClone {
    unsafe fn clone(&self) -> Self;
}

Type used for allowing unsafe cloning of internal types

Required methods

unsafe fn clone(&self) -> Self

Clone this type unsafely

Safety

Types implementing this trait perform clones under an unsafe context.

Loading content...