Trait legion::filter::EntityFilter

pub trait EntityFilter: Clone + Send {
    type ArchetypeFilter: Filter<ArchetypeFilterData<'a>> + Clone;
    type ChunksetFilter: Filter<ChunksetFilterData<'a>> + Clone;
    type ChunkFilter: Filter<ChunkFilterData<'a>> + Clone;
    fn init(&self);

    fn filters(
        &self
    ) -> (&Self::ArchetypeFilter, &Self::ChunksetFilter, &Self::ChunkFilter);

    fn into_filters(
        self
    ) -> (Self::ArchetypeFilter, Self::ChunksetFilter, Self::ChunkFilter);

    fn iter_archetype_indexes(
        &'a self, 
        storage: &'b Storage
    ) -> FilterArchIter<'b, 'a, Self::ArchetypeFilter>;

    fn iter_chunkset_indexes(
        &'a self, 
        archetype: &'b ArchetypeData
    ) -> FilterChunkIter<'b, 'a, Self::ChunksetFilter>;

    fn iter(
        &'a self, 
        storage: &'b Storage
    ) -> FilterEntityIter<'b, 'a, Self::ArchetypeFilter, Self::ChunksetFilter>;
}

A type which combines both an archetype and a chunk filter.

Associated Types

type ArchetypeFilter: Filter<ArchetypeFilterData<'a>> + Clone

type ChunksetFilter: Filter<ChunksetFilterData<'a>> + Clone

type ChunkFilter: Filter<ChunkFilterData<'a>> + Clone

Loading content...
Required methods

fn init(&self)

Initializes the entity filter for iteration.

fn filters(
&self
) -> (&Self::ArchetypeFilter, &Self::ChunksetFilter, &Self::ChunkFilter)

Gets mutable references to both inner filters.

fn into_filters(
self
) -> (Self::ArchetypeFilter, Self::ChunksetFilter, Self::ChunkFilter)

Converts self into both inner filters.

Important traits for FilterArchIter<'a, 'b, F>
impl<'a, 'b, F> Iterator for FilterArchIter<'a, 'b, F> where
F: Filter<ArchetypeFilterData<'a>>, type Item = ArchetypeIndex;

fn iter_archetype_indexes(
&'a self,
storage: &'b Storage
) -> FilterArchIter<'b, 'a, Self::ArchetypeFilter>

Gets an iterator over all matching archetype indexes.

Important traits for FilterChunkIter<'a, 'b, F>
impl<'a, 'b, F> Iterator for FilterChunkIter<'a, 'b, F> where
F: Filter<ChunksetFilterData<'a>>, type Item = SetIndex;

fn iter_chunkset_indexes(
&'a self,
archetype: &'b ArchetypeData
) -> FilterChunkIter<'b, 'a, Self::ChunksetFilter>

Gets an iterator over all matching chunkset indexes.

Important traits for FilterEntityIter<'a, 'b, Arch, Chunk>
impl<'a, 'b, Arch, Chunk> Iterator for FilterEntityIter<'a, 'b, Arch, Chunk> where
Arch: Filter<ArchetypeFilterData<'a>>,
Chunk: Filter<ChunksetFilterData<'a>>, type Item = (ArchetypeId, ChunkIndex);

fn iter(
&'a self,
storage: &'b Storage
) -> FilterEntityIter<'b, 'a, Self::ArchetypeFilter, Self::ChunksetFilter>

Gets an iterator over all matching archetypes and chunksets.

Loading content...