Trait legion::filter::Filter

pub trait Filter<T: Copy>: Send + Sync + Sized {
    type Iter: Iterator + Send + Sync;
    fn collect(&self, source: T) -> Self::Iter;

    fn is_match(&self, item: &<Self::Iter as Iterator>::Item) -> Option<bool>;

    fn matches(&mut self, source: T) -> FilterIter<Self, T> { ... }
}

A streaming iterator of bools.

Associated Types

type Iter: Iterator + Send + Sync

Loading content...
Required methods

fn collect(&self, source: T) -> Self::Iter

Pulls iterator data out of the source.

fn is_match(&self, item: &<Self::Iter as Iterator>::Item) -> Option<bool>

Determines if an element of Self::Iter matches the filter conditions.

Loading content...
Provided methods

Important traits for FilterIter<'a, F, T>
impl<'a, F: Filter<T>, T: Copy> Iterator for FilterIter<'a, F, T> type Item = bool;

fn matches(&mut self, source: T) -> FilterIter<Self, T>

Creates an iterator which yields bools for each element in the source which indicate if the element matches the filter.

Loading content...