Trait legion::iterator::FissileIterator

pub trait FissileIterator: Iterator {
    fn split(self) -> (Self, Self, usize);
}

A trait for iterators that are able to be split in roughly half. Used for splitting work among threads in parallel iterator.

Required methods

fn split(self) -> (Self, Self, usize)

Divides one iterator into two, roughly in half.

The implementation doesn't have to be precise, but the closer to the midpoint it is, the better the parallel iterator will behave.

Returns two split iterators and a number of elements left in first split. That returned size must be exact.

Loading content...