Module legion::query

Structs

Chunk
A type-safe view of a chunk of entities all of the same data layout.

ChunkDataIter

ChunkEntityIter
An iterator which iterates through all entity data in all chunks, zipped with entity ID.

ChunkViewIter
An iterator over all chunks that match a given query.

ChunkViewParIter
An iterator over all chunks that match a given query.

Query
Queries for entities within a World.

Read
Reads a single entity data component type from a chunk.

Tagged
Reads a single shared data component type in a chunk.

TryRead
Reads a single entity data component type from a chunk, if it's present.

TryWrite
Writes a single entity data component type from a chunk, if it's present.

Write
Writes to a single entity data component type from a chunk.

ZipEntities
An iterator which yields view data tuples and entity IDs from a Chunk.

Traits

DefaultFilter
A type which can construct a default entity filter.

IntoQuery
Converts a View into a Query.

View
A type which can fetch a strongly-typed view of the data contained within a chunk.