Function legion::query::maybe_changed

pub fn maybe_changed<T: Component>(
) -> EntityFilterTuple<TryComponentFilter<T>, ComponentChangedFilter<T>>

Constructs a filter which requires that the component cannot be certain to have not changed.

This check is coarse grained and should be used to reject the majority of entities which have not changed, but not all entities passed by the filter are guaranteed to have been modified.