Struct legion::query::Chunk

pub struct Chunk<'a, V: for<'b> View<'b>> { /* fields omitted */ }

A type-safe view of a chunk of entities all of the same data layout.

Methods

impl<'a, V: for<'b> View<'b>> Chunk<'a, V>

pub fn new(archetype: &'a ArchetypeData, set: usize, index: usize) -> Self

pub fn entities(&self) -> &'a [Entity]

Get a slice of all entities contained within the chunk.

pub fn iter(&mut self) -> <V as View<'a>>::Iter

Get an iterator of all data contained within the chunk.

Important traits for ZipEntities<'data, V>
impl<'data, V: View<'data>> Iterator for ZipEntities<'data, V> type Item = (Entity, <V::Iter as Iterator>::Item);

pub fn iter_entities(&mut self) -> ZipEntities<'a, V>

Get an iterator of all data and entity IDs contained within the chunk.

pub fn tag<T: Tag>(&self) -> Option<&T>

Get a tag value.

pub fn components<T: Component>(&self) -> Option<RefMap<'a, &[T]>>

Get a slice of component data.

Panics

This method performs runtime borrow checking. It will panic if any other code is concurrently writing to the data slice.

pub fn components_mut<T: Component>(&self) -> Option<RefMapMut<'a, &mut [T]>>

Get a mutable slice of component data.

Panics

This method performs runtime borrow checking. It will panic if any other code is concurrently accessing the data slice.