Trait legion::query::GroupMatcher

pub trait GroupMatcher {
    fn can_match_group() -> bool;

    fn group_components() -> Vec<ComponentTypeId>;
}

Allows a filter to determine if component optimization groups can be used to accelerate queries that use this filter.

Required methods

fn can_match_group() -> bool

Returns true if the filter may potentially match a group.

fn group_components() -> Vec<ComponentTypeId>

Returns the components that are required to be present in a group.