Trait legion::query::View

pub trait View<'a>: Sized + Send + Sync + 'static {
    type Iter: Iterator + 'a;
    fn fetch(
        archetype: &'a ArchetypeData, 
        chunk: &'a ComponentStorage, 
        chunk_index: usize
    ) -> Self::Iter;

    fn validate() -> bool;

    fn reads<T: Component>() -> bool;

    fn writes<T: Component>() -> bool;

    fn read_types() -> Vec<ComponentTypeId>;

    fn write_types() -> Vec<ComponentTypeId>;
}

A type which can fetch a strongly-typed view of the data contained within a chunk.

Associated Types

type Iter: Iterator + 'a

The iterator over the chunk data.

Loading content...
Required methods

fn fetch(
archetype: &'a ArchetypeData,
chunk: &'a ComponentStorage,
chunk_index: usize
) -> Self::Iter

Pulls data out of a chunk.

fn validate() -> bool

Validates that the view does not break any component borrowing rules.

fn reads<T: Component>() -> bool

Determines if the view reads the specified data type.

fn writes<T: Component>() -> bool

Determines if the view writes to the specified data type.

fn read_types() -> Vec<ComponentTypeId>

Returns an array of the components read by this view

fn write_types() -> Vec<ComponentTypeId>

Returns an array of the components written by this view

Loading content...