Module legion::resource

Structs

Fetch
Ergonomic wrapper type which contains a Ref type.

FetchMut
Ergonomic wrapper type which contains a RefMut type.

PreparedRead
Wrapper type for safe, lifetime-garunteed immutable access to a resource of type T'. This is the wrapper type which is provided to the closure in a System`, meaning it is only scoped to that system execution.

PreparedWrite
Wrapper type for safe, lifetime-garunteed mutable access to a resource of type T'. This is the wrapper type which is provided to the closure in a System`, meaning it is only scoped to that system execution.

ResourceTypeId
A type ID identifying a component type.

Resources
Resources container. This container stores its underlying resources in a HashMap keyed on ResourceTypeId. This means that the ID's used in this storage will not persist between recompiles.

Traits

Resource
Blanket trait for resource types.

ResourceSet
Trait which is implemented for tuples of resources and singular resources. This abstracts fetching resources to allow for ergonomic fetching.