Trait legion::resource::ResourceSet

pub trait ResourceSet: Send + Sync {
    type PreparedResources;
    fn fetch(resources: &Resources) -> Self::PreparedResources;
}

Trait which is implemented for tuples of resources and singular resources. This abstracts fetching resources to allow for ergonomic fetching.

Example:

struct TypeA(usize);
struct TypeB(usize);

use legion::prelude::*;
let mut resources = Resources::default();
resources.insert(TypeA(55));
resources.insert(TypeB(12));

{
    let (a, mut b) = <(Read<TypeA>, Write<TypeB>)>::fetch(&resources);
    assert_ne!(a.0, b.0);
    b.0 = a.0;
}

{
    let (a, b) = <(Read<TypeA>, Read<TypeB>)>::fetch(&resources);
    assert_eq!(a.0, b.0);
}

Associated Types

type PreparedResources

Loading content...
Required methods

fn fetch(resources: &Resources) -> Self::PreparedResources

Loading content...