Module legion::schedule

Structs

Builder
A factory for Schedule.

Executor
Executes a sequence of systems, potentially in parallel, and then commits their command buffers.

Schedule
A schedule of systems for execution.

Enums

ArchetypeAccess
Describes which archetypes a system declares access to.

Step
A step in a schedule.

Traits

Runnable
Trait describing a schedulable type. This is implemented by System

Schedulable
Empty trait which defines a System as schedulable by the dispatcher - this requires that the type is both Send and Sync.