Struct legion::serialize::Registry

pub struct Registry<T, S = Canon>
where
    T: TypeKey,
    S: CustomEntitySerializer + 'static,
{ /* private fields */ }

A world (de)serializer which describes how to (de)serialize the component types in a world.

The type parameter T represents the key used in the serialized output to identify each component type. The type keys used must uniquely identify each component type, and be stable between recompiles.

See the legion_typeuuid crate for an example of a type key which is stable between compiles.

Implementations

impl<T, S> Registry<T, S>
where T: TypeKey, S: CustomEntitySerializer + 'static,

pub fn new(entity_serializer: S) -> Self

Constructs a new registry.

pub fn on_unknown(&mut self, unknown: UnknownType)

Sets the behavior to use when a component type is unknown.

pub fn register<C: Component + Serialize + for<'de> Deserialize<'de>>( &mut self, mapped_type_id: T, )

Registers a component type and its key with the registry.

pub fn register_auto_mapped<C: Component + Serialize + for<'de> Deserialize<'de>>( &mut self, )
where T: AutoTypeKey<C>,

Registers a component type and its key with the registry.

pub fn as_deserialize_into_world<'a>( &'a self, world: &'a mut World, ) -> DeserializeIntoWorld<'a, Self>

Constructs a serde::DeserializeSeed which will deserialize into an existing world.

pub fn as_deserialize(&self) -> DeserializeNewWorld<'_, Self>

Constructs a serde::DeserializeSeed which will deserialize into a new world.