Struct legion::storage::PackedStorage

pub struct PackedStorage<T: Component> { /* fields omitted */ }

Stores a slice of components of type T for each archetype. Archetype slices are sorted according to the group that component T belongs to. Each slice may be packed into a single allocation to optimise for group-based access.