Struct legion::storage::TagStorage

pub struct TagStorage { /* fields omitted */ }

A vector of tag values of a single type.

Each element in the vector represents the value of tag for the chunk with the corresponding index.

Methods

impl TagStorage

pub fn element(&self) -> &TagMeta

Gets the element metadata.

pub fn len(&self) -> usize

Gets the number of tags contained within the vector.

pub fn is_empty(&self) -> bool

Determines if the vector is empty.

pub unsafe fn alloc_ptr(&mut self) -> *mut u8

Allocates uninitialized memory for a new element.

Safety

A valid element must be written into the returned address before the tag storage is next accessed.

pub unsafe fn push_raw(&mut self, ptr: *const u8)

Pushes a new tag onto the end of the vector.

Safety

Ensure the tag pointed to by ptr is representative of the tag types stored in the vec.

ptr must not point to a location already within the vector.

The value at ptr is copied into the tag vector. If the value is not Copy, then the caller must ensure that the original value is forgotten with mem::forget such that the finalizer is not called twice.

pub unsafe fn push<T>(&mut self, value: T) where
T: Tag,

Pushes a new tag onto the end of the vector.

Safety

Ensure that the type T is representative of the tag type stored in the vec.

pub unsafe fn data_raw(&self) -> (NonNull<u8>, usize, usize)

Gets a raw pointer to the start of the tag slice.

Returns a tuple containing (pointer, element_size, count).

Safety

This function returns a raw pointer with the size and length. Ensure that you do not access outside these bounds for this pointer.

Important traits for &'_ mut [u8]
impl<'_> Write for &'_ mut [u8]impl<'_> Read for &'_ [u8]

pub unsafe fn data_slice<T>(&self) -> &[T]

Gets a shared reference to the slice of tags.

Safety

Ensure that T is representative of the tag data actually stored.

Access to the tag data within the slice is runtime borrow checked. This call will panic if borrowing rules are broken.