Trait legion::storage::ArchetypeSource

pub trait ArchetypeSource {
    type Filter: LayoutFilter;

    // Required methods
    fn filter(&self) -> Self::Filter;
    fn layout(&mut self) -> EntityLayout;
}

Defines a type which can describe the layout of an archetype.

Required Associated Types

type Filter: LayoutFilter

A filter which finds existing archetypes which match the layout.

Required Methods

fn filter(&self) -> Self::Filter

Returns the archetype source’s filter.

fn layout(&mut self) -> EntityLayout

Constructs a new entity layout.