Struct legion::Resources

pub struct Resources { /* fields omitted */ }

Resources container. Shared resources stored here can be retrieved in systems.

Implementations

impl Resources

pub fn sync(&mut self) -> SyncResources<'_>

Creates an accessor to resources which are Send and Sync, which itself can be sent between threads.

pub fn contains<T: Resource>(&self) -> bool

Returns true if type T exists in the store. Otherwise, returns false.

pub fn insert<T: Resource>(&mut self, value: T)

Inserts the instance of T into the store. If the type already exists, it will be silently overwritten. If you would like to retain the instance of the resource that already exists, call remove first to retrieve it.

pub fn remove<T: Resource>(&mut self) -> Option<T>

Removes the type T from this store if it exists.

Returns

If the type T was stored, the inner instance of T is returned. Otherwise, None`.

pub fn get<T: Resource>(&self) -> Option<AtomicRef<'_, T>>

Retrieve an immutable reference to T from the store if it exists. Otherwise, return None.

Panics

Panics if the resource is already borrowed mutably.

pub fn get_mut<T: Resource>(&self) -> Option<AtomicRefMut<'_, T>>

Retrieve a mutable reference to T from the store if it exists. Otherwise, return None.

pub fn get_or_insert_with<T: Resource, F: FnOnce() -> T>(
&mut self,
f: F
) -> AtomicRef<'_, T>

Attempts to retrieve an immutable reference to T from the store. If it does not exist, the closure f is called to construct the object and it is then inserted into the store.

pub fn get_mut_or_insert_with<T: Resource, F: FnOnce() -> T>(
&mut self,
f: F
) -> AtomicRefMut<'_, T>

Attempts to retrieve a mutable reference to T from the store. If it does not exist, the closure f is called to construct the object and it is then inserted into the store.

pub fn get_or_insert<T: Resource>(&mut self, value: T) -> AtomicRef<'_, T>

Attempts to retrieve an immutable reference to T from the store. If it does not exist, the provided value is inserted and then a reference to it is returned.

pub fn get_mut_or_insert<T: Resource>(
&mut self,
value: T
) -> AtomicRefMut<'_, T>

Attempts to retrieve a mutable reference to T from the store. If it does not exist, the provided value is inserted and then a reference to it is returned.

pub fn get_or_default<T: Resource + Default>(&mut self) -> AtomicRef<'_, T>

Attempts to retrieve an immutable reference to T from the store. If it does not exist, the default constructor for T is called.

T must implement Default for this method.

pub fn get_mut_or_default<T: Resource + Default>(
&mut self
) -> AtomicRefMut<'_, T>

Attempts to retrieve a mutable reference to T from the store. If it does not exist, the default constructor for T is called.

T must implement Default for this method.

pub fn merge(&mut self, other: Resources)

Performs merging of two resource storages, which occurs during a world merge. This merge will retain any already-existant resources in the local world, while moving any new resources from the source world into this one, consuming the resources.