Module legion::system

Structs

Access
Structure used by SystemAccess for describing access to the provided T

SubWorld
Provides access to a subset of the entities of a World.

System
The concrete type which contains the system closure provided by the user. This struct should not be instantiated directly, and instead should be created using SystemBuilder.

SystemAccess
Structure describing the resource and component access conditions of the system.

SystemBuilder
The core builder of System types, which are systems within Legion. Systems are implemented as singular closures for a given system - providing queries which should be cached for that system, as well as resource access and other metadata.

SystemId

SystemQuery
A query that is usable from within a system.

Traits

QuerySet
This trait is for providing abstraction across tuples of queries for populating the type information in the system closure. This trait also provides access to the underlying query information.

SystemFn
Supertrait used for defining systems. All wrapper objects for systems implement this trait.