Struct legion::system::System

pub struct System<R, Q, F> where
    R: ResourceSet,
    Q: QuerySet,
    F: SystemFn<Resources = <R as ResourceSet>::PreparedResources, Queries = <Q as QuerySet>::Queries>,  { /* fields omitted */ }

The concrete type which contains the system closure provided by the user. This struct should not be instantiated directly, and instead should be created using SystemBuilder.

Implements Schedulable which is consumable by the StageExecutor, executing the closure.

Also handles caching of archetype information in a BitSet, as well as maintaining the provided information about what queries this system will run and, as a result, its data access.

Queries are stored generically within this struct, and the SystemQuery types are generated on each run call, wrapping the world and providing the set to the user in their closure.