Trait legion::system::QuerySet

pub trait QuerySet: Send + Sync {
    type Queries;
    fn filter_archetypes(&mut self, world: &World, archetypes: &mut BitSet);

    unsafe fn prepare(&mut self) -> Self::Queries;
}

This trait is for providing abstraction across tuples of queries for populating the type information in the system closure. This trait also provides access to the underlying query information.

Associated Types

type Queries

Loading content...
Required methods

fn filter_archetypes(&mut self, world: &World, archetypes: &mut BitSet)

Returns the archetypes accessed by this collection of queries. This allows for caching effiency and granularity for system dispatching.

unsafe fn prepare(&mut self) -> Self::Queries

Safety

prepare call doesn't respect lifetimes of self and world. The returned value cannot outlive them.

Loading content...