Struct legion::systems::CommandBuffer

pub struct CommandBuffer { /* private fields */ }

A command buffer used to queue mutable changes to the world from a system. This buffer is automatically flushed and refreshed at the beginning of every frame by Schedule. If Schedule is not used, then the user needs to manually flush it by performing CommandBuffer::flush.

Examples

Inserting an entity using the CommandBuffer:

let mut command_buffer = CommandBuffer::new(&world);
let entity = command_buffer.push(());

command_buffer.add_component(entity, Position(123.0));
command_buffer.remove(entity);

command_buffer.flush(&mut world, &mut resources);

Implementations

impl CommandBuffer

pub fn new(world: &World) -> Self

Constructs an empty command buffer.

pub fn world(&self) -> WorldId

Gets the ID of the world this command buffer belongs to.

pub fn flush(&mut self, world: &mut World, resources: &mut Resources)

Flushes this command buffer, draining all stored commands and writing them to the world.

Command flushes are performed in a FIFO manner, allowing for reliable, linear commands being executed in the order they were provided.

pub fn exec_mut<F>(&mut self, f: F) where
F: 'static + Fn(&mut World, &mut Resources) + Send + Sync,

Executes an arbitrary closure against the mutable world, allowing for queued exclusive access to the world.

pub fn push<T>(&mut self, components: T) -> Entity where
Option<T>: 'static + IntoComponentSource,
<Option<T> as IntoComponentSource>::Source: KnownLength + Send + Sync,

Queues the insertion of a single entity into the world.

pub fn extend<T>(&mut self, components: T) -> &[Entity] where
T: 'static + IntoComponentSource,
<T as IntoComponentSource>::Source: KnownLength + Send + Sync,

Queues the insertion of new entities into the world.

pub fn remove(&mut self, entity: Entity)

Queues the deletion of an entity in the command buffer.

pub fn add_component<C: Component>(&mut self, entity: Entity, component: C)

Queues the addition of a component from an entity in the command buffer.

pub fn remove_component<C: Component>(&mut self, entity: Entity)

Queues the removal of a component from an entity in the command buffer.

pub fn len(&self) -> usize

Returns the current number of commands already queued in this CommandBuffer instance.

pub fn is_empty(&self) -> bool

Returns true if this CommandBuffer is currently empty and contains no writers.