Trait legion::systems::QuerySet

pub trait QuerySet: Send + Sync {
    fn filter_archetypes(&mut self, world: &World, archetypes: &mut BitSet);
}

Provides an abstraction across tuples of queries for system closures.

Required methods

fn filter_archetypes(&mut self, world: &World, archetypes: &mut BitSet)

Evaluates the queries and records which archetypes they require access to into a bitset.