Trait legion::EntityStore

pub trait EntityStore {
    fn id(&self) -> WorldId;

    fn entry_ref(
        &self, 
        entity: Entity
    ) -> Result<EntryRef<'_>, EntityAccessError>;

    fn entry_mut(
        &mut self, 
        entity: Entity
    ) -> Result<EntryMut<'_>, EntityAccessError>;

    fn get_component_storage<V: for<'b> View<'b>>(
        &self
    ) -> Result<StorageAccessor<'_>, EntityAccessError>;
}

The EntityStore trait abstracts access to entity data as required by queries for both World and SubWorld

Required methods

fn id(&self) -> WorldId

Returns the world’s unique ID.

fn entry_ref(&self, entity: Entity) -> Result<EntryRef<'_>, EntityAccessError>

Returns an entity entry which can be used to access entity metadata and components.

fn entry_mut(
&mut self,
entity: Entity
) -> Result<EntryMut<'_>, EntityAccessError>

Returns a mutable entity entry which can be used to access entity metadata and components.

fn get_component_storage<V: for<'b> View<'b>>(
&self
) -> Result<StorageAccessor<'_>, EntityAccessError>

Returns a component storage accessor for component types declared in the specified View.