Module legion::world

Structs

ComponentTupleFilter

ComponentTupleSet
A ComponentSource which can insert tuples of components representing each entity into a world.

Universe
The Universe is a factory for creating Worlds.

World
Contains queryable collections of data associated with Entitys.

WorldId

Traits

ComponentLayout
Describes the types of a set of components attached to an entity.

ComponentSource
A set of components to be attached to one or more entities.

IntoComponentSource
An object that can be converted into a ComponentSource.

TagLayout
Describes the types of a set of tags attached to an entity.

TagSet
A set of tag values to be attached to an entity.