Struct legion::world::Duplicate

pub struct Duplicate { /* private fields */ }

A Merger which clones entities from the source world into the destination, potentially performing data transformations in the process.

Implementations

impl Duplicate

pub fn new() -> Self

Creates a new duplicate merger.

pub fn register_copy<T: Component + Copy>(&mut self)

Allows the merger to copy the given component into the destination world.

pub fn register_clone<T: Component + Clone>(&mut self)

Allows the merger to clone the given component into the destination world.

pub fn register_convert<Source: Component, Target: Component, F: FnMut(&Source) -> Target + 'static>(
&mut self,
convert: F
)

Allows the merger to clone the given component into the destination world with a custom clone function.

pub fn register_convert_raw(
&mut self,
src_type: ComponentTypeId,
dst_type: ComponentTypeId,
constructor: fn() -> Box<dyn UnknownComponentStorage>,
duplicate_fn: Box<dyn FnMut(Range<usize>, &Archetype, &dyn UnknownComponentStorage, &mut ArchetypeWriter<'_>)>
)

Allows the merger to clone the given component into the destination world with a custom clone function.