Struct legion::world::World

pub struct World {
    pub resources: Resources,
    // some fields omitted
}

Contains queryable collections of data associated with Entitys.

Fields
resources: Resources
Methods

impl World

pub fn new() -> Self

Create a new World independent of any Universe.

Entity IDs in such a world will only be unique within that world. See also Universe::create_world.

pub fn entity_channel(&mut self) -> &mut Channel<EntityEvent>

pub fn id(&self) -> WorldId

Gets the unique ID of this world within its universe.

pub fn insert<T, C>(&mut self, tags: T, components: C) -> &[Entity] where
T: TagSet + TagLayout + for<'a> Filter<ChunksetFilterData<'a>>,
C: IntoComponentSource,

Inserts new entities into the world.

Examples

Inserting entity tuples:

let tags = (model, color);
let data = vec![
    (Position(0.0), Rotation(0.0)),
    (Position(1.0), Rotation(1.0)),
    (Position(2.0), Rotation(2.0)),
];
world.insert(tags, data);

pub fn delete(&mut self, entity: Entity) -> bool

Removes the given Entity from the World.

Returns true if the entity was deleted; else false.

pub fn add_component<T: Component>(&mut self, entity: Entity, component: T)

Adds a component to an entity, or sets its value if the component is already present.

pub fn remove_component<T: Component>(&mut self, entity: Entity)

Removes a component from an entity.

pub fn add_tag<T: Tag>(&mut self, entity: Entity, tag: T)

Adds a tag to an entity, or sets its value if the tag is already present.

pub fn remove_tag<T: Tag>(&mut self, entity: Entity)

Removes a tag from an entity.

pub fn get_component<T: Component>(&self, entity: Entity) -> Option<Ref<T>>

Borrows component data for the given entity.

Returns Some(data) if the entity was found and contains the specified data. Otherwise None is returned.

pub unsafe fn get_component_mut_unchecked<T: Component>(
&self,
entity: Entity
) -> Option<RefMut<T>>

Mutably borrows entity data for the given entity.

Returns Some(data) if the entity was found and contains the specified data. Otherwise None is returned.

Safety

Accessing a component which is already being concurrently accessed elsewhere is undefined behavior.

Panics

This function may panic if any other code is currently borrowing T (such as in a query).

pub fn get_component_mut<T: Component>(
&mut self,
entity: Entity
) -> Option<RefMut<T>>

Mutably borrows entity data for the given entity.

Returns Some(data) if the entity was found and contains the specified data. Otherwise None is returned.

pub fn get_tag<T: Tag>(&self, entity: Entity) -> Option<&T>

Mutably borrows entity data for the given entity.

Returns Some(data) if the entity was found and contains the specified data. Otherwise None is returned. Gets tag data for the given entity.

Returns Some(data) if the entity was found and contains the specified data. Otherwise None is returned.

pub fn is_alive(&self, entity: Entity) -> bool

Determines if the given Entity is alive within this World.

pub fn defrag(&mut self, budget: Option<usize>)

Iteratively defragments the world's internal memory.

This compacts entities into fewer more continuous chunks.

budget describes the maximum number of entities that can be moved in one call. Subsequent calls to defrag will resume progress from the previous call.

pub fn merge(&mut self, world: World)