Trait legion::world::CloneImplResult

pub trait CloneImplResult {
    fn add_result(&mut self, src_entity: Entity, dst_entity: Entity);
}

Used along with CloneImpl, allows receiving results from a clone_from or clone_from_single call.

Required methods

fn add_result(&mut self, src_entity: Entity, dst_entity: Entity)

For every entity that is copied, this function will be called, passing the entity in the source and destination worlds

Loading content...