Trait legion::world::ComponentSource

pub trait ComponentSource: ComponentLayout {
    fn is_empty(&mut self) -> bool;

    fn len(&self) -> usize;

    fn write<T>(&mut self, entities: T, chunk: &mut ComponentStorage) -> usize
    where
        T: Iterator<Item = Entity>;
}

A set of components to be attached to one or more entities.

Required methods

fn is_empty(&mut self) -> bool

Determines if this component source has any more entity data to write.

fn len(&self) -> usize

Retrieves the nubmer of entities in this component source.

fn write<T>(&mut self, entities: T, chunk: &mut ComponentStorage) -> usize where
T: Iterator<Item = Entity>,

Writes as many components as possible into a chunk.

Loading content...