Trait legion::world::TagLayout

pub trait TagLayout {
    type Filter: Filter<ArchetypeFilterData<'a>>;
    fn get_filter(&mut self) -> &mut Self::Filter;

    fn tailor_archetype(&self, archetype: &mut ArchetypeDescription);
}

Describes the types of a set of tags attached to an entity.

Associated Types

type Filter: Filter<ArchetypeFilterData<'a>>

A filter type which filters archetypes to an exact match with this layout.

Loading content...
Required methods

fn get_filter(&mut self) -> &mut Self::Filter

Gets the archetype filter for this layout.

fn tailor_archetype(&self, archetype: &mut ArchetypeDescription)

Modifies an archetype description to include the tags described by this layout.

Loading content...