Trait legion::filter::EntityFilter

pub trait EntityFilter: Send + Clone {
    type ArchetypeFilter: for<'a> Filter<ArchetypeFilterData<'a>> + Clone;
    type ChunksetFilter: for<'a> Filter<ChunksetFilterData<'a>> + Clone;
    type ChunkFilter: for<'a> Filter<ChunkFilterData<'a>> + Clone;
    fn filters(
        &self
    ) -> (&Self::ArchetypeFilter, &Self::ChunksetFilter, &Self::ChunkFilter);

    fn into_filters(
        self
    ) -> (Self::ArchetypeFilter, Self::ChunksetFilter, Self::ChunkFilter);

    fn iter_archetype_indexes<'a, 'b>(
        &'a self, 
        storage: &'b Storage
    ) -> FilterArchIter<'b, 'a, Self::ArchetypeFilter>;

    fn iter_chunkset_indexes<'a, 'b>(
        &'a self, 
        archetype: &'b ArchetypeData
    ) -> FilterChunkIter<'b, 'a, Self::ChunksetFilter>;

    fn iter<'a, 'b>(
        &'a self, 
        storage: &'b Storage
    ) -> FilterEntityIter<'b, 'a, Self::ArchetypeFilter, Self::ChunksetFilter>;
}

A type which combines both an archetype and a chunk filter.

Associated Types

type ArchetypeFilter: for<'a> Filter<ArchetypeFilterData<'a>> + Clone

type ChunksetFilter: for<'a> Filter<ChunksetFilterData<'a>> + Clone

type ChunkFilter: for<'a> Filter<ChunkFilterData<'a>> + Clone

Loading content...
Required methods

fn filters(
&self
) -> (&Self::ArchetypeFilter, &Self::ChunksetFilter, &Self::ChunkFilter)

Gets mutable references to both inner filters.

fn into_filters(
self
) -> (Self::ArchetypeFilter, Self::ChunksetFilter, Self::ChunkFilter)

Converts self into both inner filters.

Important traits for FilterArchIter<'a, 'b, F>
impl<'a, 'b, F: Filter<ArchetypeFilterData<'a>>> Iterator for FilterArchIter<'a, 'b, F> type Item = usize;

fn iter_archetype_indexes<'a, 'b>(
&'a self,
storage: &'b Storage
) -> FilterArchIter<'b, 'a, Self::ArchetypeFilter>

Gets an iterator over all matching archetype indexes.

Important traits for FilterChunkIter<'a, 'b, F>
impl<'a, 'b, F: Filter<ChunksetFilterData<'a>>> Iterator for FilterChunkIter<'a, 'b, F> type Item = usize;

fn iter_chunkset_indexes<'a, 'b>(
&'a self,
archetype: &'b ArchetypeData
) -> FilterChunkIter<'b, 'a, Self::ChunksetFilter>

Gets an iterator over all matching chunkset indexes.

Important traits for FilterEntityIter<'a, 'b, Arch, Chunk>
impl<'a, 'b, Arch: Filter<ArchetypeFilterData<'a>>, Chunk: Filter<ChunksetFilterData<'a>>> Iterator for FilterEntityIter<'a, 'b, Arch, Chunk> type Item = (ArchetypeId, usize);

fn iter<'a, 'b>(
&'a self,
storage: &'b Storage
) -> FilterEntityIter<'b, 'a, Self::ArchetypeFilter, Self::ChunksetFilter>

Gets an iterator over all matching archetypes and chunksets.

Loading content...