Trait legion::query::View

pub trait View<'a>: 'static + Send + Sync {
    type Iter: 'a + Iterator;
    fn fetch(
        archetype: &'a ArchetypeData, 
        chunk: &'a ComponentStorage, 
        chunk_index: ChunkIndex, 
        set_index: SetIndex
    ) -> Self::Iter;

    fn validate() -> bool;

    fn validate_access(access: &ComponentAccess) -> bool;

    fn reads<T>() -> bool
    where
        T: Component;

    fn writes<T>() -> bool
    where
        T: Component;

    fn requires_permissions() -> Permissions<ComponentTypeId>;
}

A type which can fetch a strongly-typed view of the data contained within a chunk.

Associated Types

type Iter: 'a + Iterator

The iterator over the chunk data.

Loading content...
Required methods

fn fetch(
archetype: &'a ArchetypeData,
chunk: &'a ComponentStorage,
chunk_index: ChunkIndex,
set_index: SetIndex
) -> Self::Iter

Pulls data out of a chunk.

fn validate() -> bool

Validates that the view does not break any component borrowing rules.

fn validate_access(access: &ComponentAccess) -> bool

Determines if the given component access includes all permissions required by the view.

fn reads<T>() -> bool where
T: Component,

Determines if the view reads the specified data type.

fn writes<T>() -> bool where
T: Component,

Determines if the view writes to the specified data type.

fn requires_permissions() -> Permissions<ComponentTypeId>

Returns the set of permissions required by the view.

Loading content...